requirements.txt
fastapi
uvicorn
pydantic
SQLAlchemy
mysql-connector-python

